Move server/util to p2p/util
