+++ p2p/util.js
